yamlall-standard-multiself/requirements.txt
pyyaml
numpy
flake8
pytest
